+++ src/kard/core/client_wrapper.py
@@ -27,12 +27,12 @@ class BaseClientWrapper:
         import platform
 
         headers: typing.Dict[str, str] = {
-            "User-Agent": "kard-financial-sdk/8.2.0",
+            "User-Agent": "kard-financial-sdk/8.3.0",
             "X-Fern-Language": "Python",
             "X-Fern-Runtime": f"python/{platform.python_version()}",
             "X-Fern-Platform": f"{platform.system().lower()}/{platform.release()}",
             "X-Fern-SDK-Name": "kard-financial-sdk",
-            "X-Fern-SDK-Version": "8.2.0",
+            "X-Fern-SDK-Version": "8.3.0",
             **(self.get_custom_headers() or {}),
         }
         token = self._get_token()